--- src/IlluminaSampleSheet/V1/NovaSeqXpSampleSheet.php
@@ -1,20 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace MLL\Utils\IlluminaSampleSheet\V1;
-
-use MLL\Utils\IlluminaSampleSheet\SampleSheet;
-
-class NovaSeqXpSampleSheet extends SampleSheet
-{
-    public function __construct(
-        NovaSeqHeaderSection $header,
-        ReadsSection $reads,
-        SettingsSection $settings,
-        DataSection $data
-    ) {
-        $this->addSection($header);
-        $this->addSection($reads);
-        $this->addSection($settings);
-        $this->addSection($data);
-    }
-}
