+++ tests/IlluminaSampleSheet/V1/NovaSeqSampleSheetTest.php
@@ -2,8 +2,11 @@
 
 namespace MLL\Utils\Tests\IlluminaSampleSheet\V1;
 
-use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSample;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqDataSection;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection;
 use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSampleSheet;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqXpSample;
+use MLL\Utils\IlluminaSampleSheet\V1\ReadsSection;
 use MLL\Utils\IlluminaSampleSheet\V1\SettingsSection;
 use PHPUnit\Framework\TestCase;
 
@@ -11,10 +14,10 @@ class NovaSeqSampleSheetTest extends TestCase
 {
     public function testNovaSeqSampleSheetAddsSectionsOnConstruction(): void
     {
-        $headerSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection::class);
-        $readsSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\ReadsSection::class);
-        $settingsSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\SettingsSection::class);
-        $dataSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\DataSection::class);
+        $headerSection = $this->createMock(NovaSeqHeaderSection::class);
+        $readsSection = $this->createMock(ReadsSection::class);
+        $settingsSection = $this->createMock(SettingsSection::class);
+        $dataSection = $this->createMock(NovaSeqDataSection::class);
 
         $novaSeqSampleSheet = new NovaSeqSampleSheet($headerSection, $readsSection, $settingsSection, $dataSection);
 
@@ -26,7 +29,7 @@ class NovaSeqSampleSheetTest extends TestCase
 
     public function testNovaSeqStandardSampleSheetToStringReturnsExpectedResult(): void
     {
-        $headerSection = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection(
+        $headerSection = new NovaSeqHeaderSection(
             '4',
             'DonalDuck',
             'MyExperiment',
@@ -38,16 +41,16 @@ class NovaSeqSampleSheetTest extends TestCase
             'MyChemistry',
         );
 
-        $readsSection = new \MLL\Utils\IlluminaSampleSheet\V1\ReadsSection(101, 101);
+        $readsSection = new ReadsSection(101, 101);
 
-        $dataSection = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqDataSection();
+        $dataSection = new NovaSeqDataSection();
 
-        $dataSection->addSample(new NovaSeqSample('1', 'Sample-001-M001', 'RunXXXX-PLATE', '', 'UDP0090', 'TCAGGCTT', 'UDP0090', 'ATCATGCG', 'RunXXXX-PROJECT', 'description'));
-        $dataSection->addSample(new NovaSeqSample('2', 'Sample-002-M002', 'RunXXXX-PLATE', '', 'UDP0091', 'CCTTGTAG', 'UDP0091', 'CCTTGGAA', 'RunXXXX-PROJECT', 'description'));
-        $dataSection->addSample(new NovaSeqSample('3', 'Sample-003-M003', 'RunXXXX-PLATE', '', 'UDP0092', 'GAACATCG', 'UDP0092', 'TCGACAAG', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqXpSample(1, '1', 'Sample-001-M001', 'RunXXXX-PLATE', '', 'UDP0090', 'TCAGGCTT', 'UDP0090', 'ATCATGCG', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqXpSample(1, '2', 'Sample-002-M002', 'RunXXXX-PLATE', '', 'UDP0091', 'CCTTGTAG', 'UDP0091', 'CCTTGGAA', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqXpSample(1, '3', 'Sample-003-M003', 'RunXXXX-PLATE', '', 'UDP0092', 'GAACATCG', 'UDP0092', 'TCGACAAG', 'RunXXXX-PROJECT', 'description'));
 
         $settings = new SettingsSection('AGATCGGAAGAGCACACGTCTGAACTCCAGTCA', 'AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT');
-        $miSeqSampleSheet = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSampleSheet($headerSection, $readsSection, $settings, $dataSection);
+        $miSeqSampleSheet = new NovaSeqSampleSheet($headerSection, $readsSection, $settings, $dataSection);
 
         $expected = '[Header]
 IEMFileVersion,4
@@ -70,6 +73,56 @@ Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,Index,I5_Index_ID,Ind
 1,Sample-001-M001,RunXXXX-PLATE,,UDP0090,TCAGGCTT,UDP0090,ATCATGCG,RunXXXX-PROJECT,description
 2,Sample-002-M002,RunXXXX-PLATE,,UDP0091,CCTTGTAG,UDP0091,CCTTGGAA,RunXXXX-PROJECT,description
 3,Sample-003-M003,RunXXXX-PLATE,,UDP0092,GAACATCG,UDP0092,TCGACAAG,RunXXXX-PROJECT,description
+';
+        self::assertSame($expected, $miSeqSampleSheet->toString());
+    }
+
+    public function testNovaSeqXpSampleSheetWithLanesToStringReturnsExpectedResult(): void
+    {
+        $headerSection = new NovaSeqHeaderSection(
+            '4',
+            'DonalDuck',
+            'MyExperiment',
+            '19.04.2024',
+            'MyWorkflow',
+            'MyApplication',
+            'MyAssay',
+            'MyDescription',
+            'MyChemistry',
+        );
+
+        $readsSection = new ReadsSection(101, 101);
+
+        $dataSection = new NovaSeqDataSection();
+
+        $dataSection->addSample(new NovaSeqXpSample(2, '1', 'Sample-001-M001', 'RunXXXX-PLATE', '', 'UDP0090', 'TCAGGCTT', 'UDP0090', 'ATCATGCG', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqXpSample(1, '2', 'Sample-002-M002', 'RunXXXX-PLATE', '', 'UDP0091', 'CCTTGTAG', 'UDP0091', 'CCTTGGAA', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqXpSample(4, '3', 'Sample-003-M003', 'RunXXXX-PLATE', '', 'UDP0092', 'GAACATCG', 'UDP0092', 'TCGACAAG', 'RunXXXX-PROJECT', 'description'));
+
+        $settings = new SettingsSection('AGATCGGAAGAGCACACGTCTGAACTCCAGTCA', 'AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT');
+        $miSeqSampleSheet = new NovaSeqSampleSheet($headerSection, $readsSection, $settings, $dataSection);
+
+        $expected = '[Header]
+IEMFileVersion,4
+Investigator Name,DonalDuck
+Experiment Name,MyExperiment
+Date,19.04.2024
+Workflow,MyWorkflow
+Application,MyApplication
+Assay,MyAssay
+Description,MyDescription
+Chemistry,MyChemistry
+[Reads]
+101
+101
+[Settings]
+Adapter,AGATCGGAAGAGCACACGTCTGAACTCCAGTCA
+AdapterRead2,AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT
+[Data]
+Lane,Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,Index,I5_Index_ID,Index2,Sample_Project,Description
+2,1,Sample-001-M001,RunXXXX-PLATE,,UDP0090,TCAGGCTT,UDP0090,ATCATGCG,RunXXXX-PROJECT,description
+1,2,Sample-002-M002,RunXXXX-PLATE,,UDP0091,CCTTGTAG,UDP0091,CCTTGGAA,RunXXXX-PROJECT,description
+4,3,Sample-003-M003,RunXXXX-PLATE,,UDP0092,GAACATCG,UDP0092,TCGACAAG,RunXXXX-PROJECT,description
 ';
         self::assertSame($expected, $miSeqSampleSheet->toString());
     }
